requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
jinja2==3.1.6
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.0
PyJWT==2.9.0
email-validator==2.2.0
bleach==6.2.0
pytest==8.3.5
itsdangerous==2.2.0
